# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Enhanced Authentication API",
    "description": "The deployed API is responsible for enhanced authentication.",
    "contact": {
      "email": "chandangowda.kr2@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8010/api/v1/enhanced-authentication"
    }
  ],
...
  "paths": {
...
    "/user/publicProfiles": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Lists all the public profiles.",
        "description": "Lists all the public profiles.",
        "operationId": "listPublicProfiles",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "theUser"
          },
          "email": {
            "type": "string",
            "example": "john@email.com"
          },
          "publicProfile": {
            "type": "boolean",
            "example": true
          },
          "role": {
            "type": "string",
            "description": "Admin or User",
            "example": "admin"
          }
        },
        "xml": {
          "name": "User"
        }
      }
    },
...
  }
}